Decouple AuthController from AuthTokenRepository
