+++ app-backend/src/main/kotlin/org/example/appbackend/service/AuthService.kt
@@ -1,0 +1,7 @@
+package org.example.appbackend.service
+
+interface AuthService {
+
+    fun createToken(userId : Int): String
+    fun deleteToken(token: String)
+}
